Adding a migrate method and adding the logic to convert status channels
This commit introduce a migrate procedure to be tested, and in addition, this include a json conversion of status channel.

It was stored like a map to make the access easy but it works out of the graphql convention. For this reason, the json payload now contains a list of status channel, and the final user need to manage the access to this datastructure.

Missing step are:
- [ ] Adding metric version in the code, this need to be in the payload to manage the conversion.
- [ ] testing the migration options to migrate the actual map of status channels with a list of status channels.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>